+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/theme/ChatTheme.kt
@@ -15,6 +15,7 @@ import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
 import io.getstream.chat.android.compose.ui.attachments.StreamAttachmentFactories
 import io.getstream.chat.android.compose.ui.util.ChannelNameFormatter
 import io.getstream.chat.android.compose.ui.util.DefaultReactionTypes
+import io.getstream.chat.android.compose.ui.util.MessageAlignmentProvider
 import io.getstream.chat.android.compose.ui.util.MessagePreviewFormatter
 import io.getstream.chat.android.compose.ui.util.StreamCoilImageLoader
 
@@ -39,17 +40,17 @@ private val LocalAttachmentFactories = compositionLocalOf<List<AttachmentFactory
 private val LocalReactionTypes = compositionLocalOf<Map<String, Int>> {
     error("No reactions provided! Make sure to wrap all usages of Stream components in a ChatTheme.")
 }
-
 private val LocalDateFormatter = compositionLocalOf<DateFormatter> {
     error("No DateFormatter provided! Make sure to wrap all usages of Stream components in a ChatTheme.")
 }
-
 private val LocalChannelNameFormatter = compositionLocalOf<ChannelNameFormatter> {
     error("No ChannelNameFormatter provided! Make sure to wrap all usages of Stream components in a ChatTheme.")
 }
-
 private val LocalMessagePreviewFormatter = compositionLocalOf<MessagePreviewFormatter> {
     error("No MessagePreviewFormatter provided! Make sure to wrap all usages of Stream components in a ChatTheme.")
+}
+private val LocalMessageAlignmentProvider = compositionLocalOf<MessageAlignmentProvider> {
+    error("No MessageAlignmentProvider provided! Make sure to wrap all usages of Stream components in a ChatTheme.")
 }
 
 /**
@@ -66,6 +67,7 @@ private val LocalMessagePreviewFormatter = compositionLocalOf<MessagePreviewForm
  * @param dateFormatter [DateFormatter] used throughout the app for date and time information.
  * @param channelNameFormatter [ChannelNameFormatter] used throughout the app for channel names.
  * @param messagePreviewFormatter [MessagePreviewFormatter] used to generate a string preview for the given message.
+ * @param messageAlignmentProvider [MessageAlignmentProvider] used to provide message alignment for the given message.
  * @param content The content shown within the theme wrapper.
  */
 @Composable
@@ -80,6 +82,7 @@ public fun ChatTheme(
     dateFormatter: DateFormatter = DateFormatter.from(LocalContext.current),
     channelNameFormatter: ChannelNameFormatter = ChannelNameFormatter.defaultFormatter(LocalContext.current),
     messagePreviewFormatter: MessagePreviewFormatter = MessagePreviewFormatter.defaultFormatter(LocalContext.current),
+    messageAlignmentProvider: MessageAlignmentProvider = MessageAlignmentProvider.defaultMessageAlignmentProvider(),
     content: @Composable () -> Unit,
 ) {
     LaunchedEffect(Unit) {
@@ -96,7 +99,8 @@ public fun ChatTheme(
         LocalDateFormatter provides dateFormatter,
         LocalChannelNameFormatter provides channelNameFormatter,
         LocalMessagePreviewFormatter provides messagePreviewFormatter,
-        LocalImageLoader provides StreamCoilImageLoader.imageLoader(LocalContext.current)
+        LocalImageLoader provides StreamCoilImageLoader.imageLoader(LocalContext.current),
+        LocalMessageAlignmentProvider provides MessageAlignmentProvider.defaultMessageAlignmentProvider()
     ) {
         content()
     }
@@ -152,4 +156,9 @@ public object ChatTheme {
         @Composable
         @ReadOnlyComposable
         get() = LocalMessagePreviewFormatter.current
+
+    public val messageAlignmentProvider: MessageAlignmentProvider
+        @Composable
+        @ReadOnlyComposable
+        get() = LocalMessageAlignmentProvider.current
 }